Update to account for the MID-GAME SHUFFLE
The first game I played actually using the tool, I saw that the
cards are actually shuffled when the player chooses to stand.

I have no idea how I missed this when writing the page initially.
I'm thrilled that it's a one-line change to update the tool.

Well, one line of code plus the one-line comment change :)